+++ Areas/Master/Controllers/AccountSetupController.cs
@@ -61,7 +61,7 @@ namespace AEMSWEB.Areas.Master.Controllers
             if (pageNumber < 1 || pageSize < 1)
                 return Json(new { success = false, message = "Invalid page parameters" });
 
-            var validationResult = ValidateCompanyAndUserId(companyId, out short companyIdShort, out short? parsedUserId);
+            var validationResult = ValidateCompanyAndUserId(companyId, out byte companyIdShort, out short? parsedUserId);
             if (validationResult != null) return validationResult;
 
             try
@@ -83,7 +83,7 @@ namespace AEMSWEB.Areas.Master.Controllers
             if (accSetupId <= 0)
                 return Json(new { success = false, message = "Invalid Account Setup ID" });
 
-            var validationResult = ValidateCompanyAndUserId(companyId, out short companyIdShort, out short? parsedUserId);
+            var validationResult = ValidateCompanyAndUserId(companyId, out byte companyIdShort, out short? parsedUserId);
             if (validationResult != null) return validationResult;
 
             try
@@ -106,7 +106,7 @@ namespace AEMSWEB.Areas.Master.Controllers
             if (model == null || !ModelState.IsValid)
                 return Json(new { success = false, message = "Invalid request data" });
 
-            var validationResult = ValidateCompanyAndUserId(model.companyId, out short companyIdShort, out short? parsedUserId);
+            var validationResult = ValidateCompanyAndUserId(model.companyId, out byte companyIdShort, out short? parsedUserId);
             if (validationResult != null) return validationResult;
 
             try
@@ -142,7 +142,7 @@ namespace AEMSWEB.Areas.Master.Controllers
             if (accSetupId <= 0)
                 return Json(new { success = false, message = "Invalid Account Setup ID" });
 
-            var validationResult = ValidateCompanyAndUserId(companyId, out short companyIdShort, out short? parsedUserId);
+            var validationResult = ValidateCompanyAndUserId(companyId, out byte companyIdShort, out short? parsedUserId);
             if (validationResult != null) return validationResult;
 
             var permissions = await HasPermission(companyIdShort, parsedUserId.Value,
@@ -173,7 +173,7 @@ namespace AEMSWEB.Areas.Master.Controllers
             if (pageNumber < 1 || pageSize < 1)
                 return Json(new { success = false, message = "Invalid page parameters" });
 
-            var validationResult = ValidateCompanyAndUserId(companyId, out short companyIdShort, out short? parsedUserId);
+            var validationResult = ValidateCompanyAndUserId(companyId, out byte companyIdShort, out short? parsedUserId);
             if (validationResult != null) return validationResult;
 
             try
@@ -195,7 +195,7 @@ namespace AEMSWEB.Areas.Master.Controllers
             if (accSetupId <= 0)
                 return Json(new { success = false, message = "Invalid Account Setup Detail ID" });
 
-            var validationResult = ValidateCompanyAndUserId(companyId, out short companyIdShort, out short? parsedUserId);
+            var validationResult = ValidateCompanyAndUserId(companyId, out byte companyIdShort, out short? parsedUserId);
             if (validationResult != null) return validationResult;
 
             try
@@ -218,7 +218,7 @@ namespace AEMSWEB.Areas.Master.Controllers
             if (model == null || !ModelState.IsValid)
                 return Json(new { success = false, message = "Invalid request data" });
 
-            var validationResult = ValidateCompanyAndUserId(model.companyId, out short companyIdShort, out short? parsedUserId);
+            var validationResult = ValidateCompanyAndUserId(model.companyId, out byte companyIdShort, out short? parsedUserId);
             if (validationResult != null) return validationResult;
 
             try
@@ -253,7 +253,7 @@ namespace AEMSWEB.Areas.Master.Controllers
             if (accSetupId <= 0)
                 return Json(new { success = false, message = "Invalid Account Setup Detail ID" });
 
-            var validationResult = ValidateCompanyAndUserId(companyId, out short companyIdShort, out short? parsedUserId);
+            var validationResult = ValidateCompanyAndUserId(companyId, out byte companyIdShort, out short? parsedUserId);
             if (validationResult != null) return validationResult;
 
             var permissions = await HasPermission(companyIdShort, parsedUserId.Value,
@@ -284,7 +284,7 @@ namespace AEMSWEB.Areas.Master.Controllers
             if (pageNumber < 1 || pageSize < 1)
                 return Json(new { success = false, message = "Invalid page parameters" });
 
-            var validationResult = ValidateCompanyAndUserId(companyId, out short companyIdShort, out short? parsedUserId);
+            var validationResult = ValidateCompanyAndUserId(companyId, out byte companyIdShort, out short? parsedUserId);
             if (validationResult != null) return validationResult;
 
             try
@@ -306,7 +306,7 @@ namespace AEMSWEB.Areas.Master.Controllers
             if (accSetupCategoryId <= 0)
                 return Json(new { success = false, message = "Invalid Account Setup Category ID" });
 
-            var validationResult = ValidateCompanyAndUserId(companyId, out short companyIdShort, out short? parsedUserId);
+            var validationResult = ValidateCompanyAndUserId(companyId, out byte companyIdShort, out short? parsedUserId);
             if (validationResult != null) return validationResult;
 
             try
@@ -329,7 +329,7 @@ namespace AEMSWEB.Areas.Master.Controllers
             if (model == null || !ModelState.IsValid)
                 return Json(new { success = false, message = "Invalid request data" });
 
-            var validationResult = ValidateCompanyAndUserId(model.companyId, out short companyIdShort, out short? parsedUserId);
+            var validationResult = ValidateCompanyAndUserId(model.companyId, out byte companyIdShort, out short? parsedUserId);
             if (validationResult != null) return validationResult;
 
             try
@@ -363,7 +363,7 @@ namespace AEMSWEB.Areas.Master.Controllers
             if (accSetupCategoryId <= 0)
                 return Json(new { success = false, message = "Invalid Account Setup Category ID" });
 
-            var validationResult = ValidateCompanyAndUserId(companyId, out short companyIdShort, out short? parsedUserId);
+            var validationResult = ValidateCompanyAndUserId(companyId, out byte companyIdShort, out short? parsedUserId);
             if (validationResult != null) return validationResult;
 
             var permissions = await HasPermission(companyIdShort, parsedUserId.Value,
