+++ Areas/Master/Controllers/BankController.cs
@@ -61,7 +61,7 @@ namespace AEMSWEB.Areas.Master.Controllers
             if (pageNumber < 1 || pageSize < 1)
                 return Json(new { success = false, message = "Invalid page parameters" });
 
-            var validationResult = ValidateCompanyAndUserId(companyId, out short companyIdShort, out short? parsedUserId);
+            var validationResult = ValidateCompanyAndUserId(companyId, out byte companyIdShort, out short? parsedUserId);
             if (validationResult != null) return validationResult;
 
             try
@@ -83,7 +83,7 @@ namespace AEMSWEB.Areas.Master.Controllers
             if (customerId <= 0)
                 return Json(new { success = false, message = "Invalid Bank ID" });
 
-            var validationResult = ValidateCompanyAndUserId(companyId, out short companyIdShort, out short? parsedUserId);
+            var validationResult = ValidateCompanyAndUserId(companyId, out byte companyIdShort, out short? parsedUserId);
             if (validationResult != null) return validationResult;
 
             try
@@ -106,7 +106,7 @@ namespace AEMSWEB.Areas.Master.Controllers
             if (model == null || !ModelState.IsValid)
                 return Json(new { success = false, message = "Invalid request data" });
 
-            var validationResult = ValidateCompanyAndUserId(model.companyId, out short companyIdShort, out short? parsedUserId);
+            var validationResult = ValidateCompanyAndUserId(model.companyId, out byte companyIdShort, out short? parsedUserId);
             if (validationResult != null) return validationResult;
 
             try
@@ -138,7 +138,7 @@ namespace AEMSWEB.Areas.Master.Controllers
         [HttpDelete]
         public async Task<IActionResult> Delete(short BankId, string companyId)
         {
-            var validationResult = ValidateCompanyAndUserId(companyId, out short companyIdShort, out short? parsedUserId);
+            var validationResult = ValidateCompanyAndUserId(companyId, out byte companyIdShort, out short? parsedUserId);
             if (validationResult != null) return validationResult;
 
             var permissions = await HasPermission(companyIdShort, parsedUserId.Value,
@@ -166,7 +166,7 @@ namespace AEMSWEB.Areas.Master.Controllers
         [HttpGet]
         public async Task<JsonResult> ListContacts(int BankId, string companyId)
         {
-            var validationResult = ValidateCompanyAndUserId(companyId, out short companyIdShort, out short? parsedUserId);
+            var validationResult = ValidateCompanyAndUserId(companyId, out byte companyIdShort, out short? parsedUserId);
             if (validationResult != null) return validationResult;
 
             try
@@ -188,7 +188,7 @@ namespace AEMSWEB.Areas.Master.Controllers
             if (contactId <= 0)
                 return Json(new { success = false, message = "Invalid Contact ID" });
 
-            var validationResult = ValidateCompanyAndUserId(companyId, out short companyIdShort, out short? parsedUserId);
+            var validationResult = ValidateCompanyAndUserId(companyId, out byte companyIdShort, out short? parsedUserId);
             if (validationResult != null) return validationResult;
 
             try
@@ -211,7 +211,7 @@ namespace AEMSWEB.Areas.Master.Controllers
             if (model == null || !ModelState.IsValid)
                 return Json(new { success = false, message = "Invalid request data" });
 
-            var validationResult = ValidateCompanyAndUserId(model.companyId, out short companyIdShort, out short? parsedUserId);
+            var validationResult = ValidateCompanyAndUserId(model.companyId, out byte companyIdShort, out short? parsedUserId);
             if (validationResult != null) return validationResult;
 
             try
@@ -254,7 +254,7 @@ namespace AEMSWEB.Areas.Master.Controllers
             if (contactId <= 0)
                 return Json(new { success = false, message = "Invalid Contact ID" });
 
-            var validationResult = ValidateCompanyAndUserId(companyId, out short companyIdShort, out short? parsedUserId);
+            var validationResult = ValidateCompanyAndUserId(companyId, out byte companyIdShort, out short? parsedUserId);
             if (validationResult != null) return validationResult;
 
             var permissions = await HasPermission(companyIdShort, parsedUserId.Value,
@@ -282,7 +282,7 @@ namespace AEMSWEB.Areas.Master.Controllers
         [HttpGet]
         public async Task<JsonResult> ListAddresses(int BankId, string companyId)
         {
-            var validationResult = ValidateCompanyAndUserId(companyId, out short companyIdShort, out short? parsedUserId);
+            var validationResult = ValidateCompanyAndUserId(companyId, out byte companyIdShort, out short? parsedUserId);
             if (validationResult != null) return validationResult;
 
             try
@@ -304,7 +304,7 @@ namespace AEMSWEB.Areas.Master.Controllers
             if (addressId <= 0)
                 return Json(new { success = false, message = "Invalid Address ID" });
 
-            var validationResult = ValidateCompanyAndUserId(companyId.ToString(), out short companyIdShort, out short? parsedUserId);
+            var validationResult = ValidateCompanyAndUserId(companyId.ToString(), out byte companyIdShort, out short? parsedUserId);
             if (validationResult != null) return validationResult;
 
             try
@@ -327,7 +327,7 @@ namespace AEMSWEB.Areas.Master.Controllers
             if (model == null || !ModelState.IsValid)
                 return Json(new { success = false, message = "Invalid request data" });
 
-            var validationResult = ValidateCompanyAndUserId(model.companyId, out short companyIdShort, out short? parsedUserId);
+            var validationResult = ValidateCompanyAndUserId(model.companyId, out byte companyIdShort, out short? parsedUserId);
             if (validationResult != null) return validationResult;
 
             try
@@ -373,7 +373,7 @@ namespace AEMSWEB.Areas.Master.Controllers
             if (contactId <= 0)
                 return Json(new { success = false, message = "Invalid Address ID" });
 
-            var validationResult = ValidateCompanyAndUserId(companyId, out short companyIdShort, out short? parsedUserId);
+            var validationResult = ValidateCompanyAndUserId(companyId, out byte companyIdShort, out short? parsedUserId);
             if (validationResult != null) return validationResult;
 
             var permissions = await HasPermission(companyIdShort, parsedUserId.Value,
